refactor: Organize sidebar components and update dashboard styling
- Moved sidebar-related components (AppSidebar, NavMain, NavProjects, NavUser, TeamSwitcher) into a new `components/sidebar/` directory for improved project structure.
- Renamed internal navigation components within AppSidebar to use a `Sidebar` prefix for consistency (e.g., `NavMain` to `SidebarMain`).
- Updated import paths in `app/(authenticated)/layout.tsx` and `components/sidebar/app-sidebar.tsx` to reflect the new locations and names.
- Adjusted styling in `app/(authenticated)/(user)/dashboard/page.tsx`: reduced `gap` and removed `mb-4` from the heading for a more compact layout.
- Added `cursor-pointer` style to `SidebarTrigger` in `components/ui/sidebar.tsx` for better user experience.
