+++ components/sidebar/sidebar-main.tsx
@@ -18,7 +18,7 @@ import {
   SidebarMenuSubItem,
 } from "@/components/ui/sidebar"
 
-export function NavMain({
+export function SidebarMain({
   items,
 }: {
   items: {
