+++ components/sidebar/sidebar-projects.tsx
@@ -25,7 +25,7 @@ import {
   useSidebar,
 } from "@/components/ui/sidebar"
 
-export function NavProjects({
+export function SidebarProjects({
   projects,
 }: {
   projects: {
