+++ components/sidebar/sidebar-user.tsx
@@ -30,7 +30,7 @@ import {
   useSidebar,
 } from "@/components/ui/sidebar"
 
-export function NavUser({
+export function SidebarUser({
   user,
 }: {
   user: {
